🔧 Implement message storing in queues

+++ internal/config/queue.go
@@ -3,8 +3,8 @@ package config
 // Queue конфигурация очереди сообщений.
 type Queue struct {
 	// MaxSize максимальный размер очереди.
-	MaxSize int64 `mapstructure:"max_size"`
+	MaxSize int `mapstructure:"max_size"`
 
 	// MaxConsumers максимальное количество подписчиков.
-	MaxConsumers int64 `mapstructure:"max_consumers"`
+	MaxConsumers int `mapstructure:"max_consumers"`
 }
